requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.0.8
django-tastypie==0.14.6
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-mimeparse==1.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
